requirements.txt
gradio-client==0.1.1
Pillow==9.1.1
gradio-client==0.1.1
Pillow==9.1.1
torch==1.13.1
torchvision==0.14.1
